Break up Helpers.cs
