+++ src/Bicep.Wasm/LanguageHelpers/Position.cs
@@ -1,0 +1,15 @@
+﻿namespace Bicep.Wasm.LanguageHelpers
+{
+    public class Position
+    {
+        public Position(int line, int character)
+        {
+            Line = line;
+            Character = character;
+        }
+
+        public int Line { get; }
+
+        public int Character { get; }
+    }
+}
