+++ src/Bicep.Wasm/LanguageHelpers/Range.cs
@@ -1,0 +1,20 @@
+﻿namespace Bicep.Wasm.LanguageHelpers
+{
+    public class Range
+    {
+        public Range(Position start, Position end)
+        {
+            Start = start;
+            End = end;
+        }
+
+        public Range()
+            : this(new Position(0, 0), new Position(0, 0))
+        {
+        }
+
+        public Position Start { get; set; }
+
+        public Position End { get; set; }
+    }
+}
