+++ src/Bicep.Wasm/LanguageHelpers/SemanticTokenType.cs
@@ -1,0 +1,29 @@
+﻿namespace Bicep.Wasm.LanguageHelpers
+{
+    public enum SemanticTokenType
+    {
+        Comment,
+        EnumMember,
+        Event,
+        Modifier,
+        Label,
+        Parameter,
+        Variable,
+        Property,
+        Member,
+        Function,
+        TypeParameter,
+        Macro,
+        Interface,
+        Enum,
+        String,
+        Number,
+        Regexp,
+        Operator,
+        Keyword,
+        Type,
+        Struct,
+        Class,
+        Namespace,
+    }
+}
